feat(protocol): Scale up damping factor and flatten curve (#13809)

+++ packages/protocol/contracts/L1/TaikoData.sol
@@ -27,7 +27,7 @@ library TaikoData {
         uint64 maxEthDepositsPerBlock;
         uint96 maxEthDepositAmount;
         uint96 minEthDepositAmount;
-        uint8 adjustmentQuotient;
+        uint16 adjustmentQuotient;
         bool relaySignalRoot;
     }
